+++ nn_lib/mdl/layers/__init__.py
@@ -1,0 +1,1 @@
+from nn_lib.mdl.layers.linear import Linear
