+++ nn_lib/mdl/loss_functions/__init__.py
@@ -1,0 +1,2 @@
+from nn_lib.mdl.loss_functions.loss import Loss
+from nn_lib.mdl.loss_functions.bce_loss import BCELoss
